requirements.txt
Flask>=2.0.0
Pillow>=8.0.0
smbus2>=0.4.3
luma.core>=1.0.0
luma.oled>=1.0.0
build>=1.2.2
colorama>=0.4.6